4. Sınıf/Güz Dönemi/Ağ Programlama/Proje-Tasarım/G221210045-G221210383/Secure-E2EE-Chat-App-main/requirements.txt
psycopg2-binary
asyncio